Add passwordsource-env.patch to allow passing passwords through ENV
If we're passing arguments on the command line others might pick it up
(either through the /proc interface or through execve() logging or
whatever).

The --passwordSource env setting allows you to sed SEDUTIL_PASSWORD=x
instead. You'll need to place a '-' where the password used to be in the
command line. For password changes you'll use SEDUTIL_NEWPASSWORD=y for
the second password.